refactor: move tests around
