+++ packages/oas-examples/3.0/json/server-variables.json
@@ -168,6 +168,70 @@
           }
         }
       }
+    },
+    "/relative-path-server": {
+      "servers": [{ "url": "/v2" }],
+      "get": {
+        "summary": "Relative path-level server",
+        "tags": ["Path"],
+        "responses": {
+          "200": {
+            "description": "OK"
+          }
+        }
+      }
+    },
+    "/relative-operation-server": {
+      "get": {
+        "summary": "Relative operation-level server",
+        "tags": ["Operation"],
+        "servers": [{ "url": "/v3" }],
+        "responses": {
+          "200": {
+            "description": "OK"
+          }
+        }
+      }
+    },
+    "/path-item-ref-server": {
+      "$ref": "#/paths/~1path-item-server-source"
+    },
+    "/path-item-server-source": {
+      "servers": [{ "url": "https://path-item-ref.example.com" }],
+      "get": {
+        "summary": "Path item ref server source",
+        "tags": ["Path"],
+        "responses": {
+          "200": {
+            "description": "OK"
+          }
+        }
+      }
+    },
+    "/empty-operation-servers": {
+      "servers": [{ "url": "https://empty-operation-path.example.com" }],
+      "get": {
+        "summary": "Empty operation-level servers",
+        "tags": ["Operation"],
+        "servers": [],
+        "responses": {
+          "200": {
+            "description": "OK"
+          }
+        }
+      }
+    },
+    "/empty-path-item-servers": {
+      "servers": [],
+      "get": {
+        "summary": "Empty path-level servers",
+        "tags": ["Path"],
+        "responses": {
+          "200": {
+            "description": "OK"
+          }
+        }
+      }
     }
   }
 }
