+++ packages/oas/test/index.test.ts
@@ -13,7 +13,6 @@ import circularSpec from './__datasets__/circular.json' with { type: 'json' };
 import complexNestingSpec from './__datasets__/complex-nesting.json' with { type: 'json' };
 import invalidComponentSchemaNamesSpec from './__datasets__/invalid-component-schema-names.json' with { type: 'json' };
 import multipleSecuritiesSpec from './__datasets__/multiple-securities.json' with { type: 'json' };
-import operationServerPrecedenceSpec from './__datasets__/operation-server-precedence.json' with { type: 'json' };
 import orderedTagsSpec from './__datasets__/ordered-tags.json' with { type: 'json' };
 import pathMatchingQuirksSpec from './__datasets__/path-matching-quirks.json' with { type: 'json' };
 import pathVariableQuirksSpec from './__datasets__/path-variable-quirks.json' with { type: 'json' };
@@ -625,144 +624,6 @@ describe('Oas', () => {
           paths: {},
         }).variables(10),
       ).toStrictEqual({});
-    });
-  });
-
-  describe('Operation server helpers', () => {
-    it('should fall back to root-level servers', () => {
-      const oas = Oas.init(operationServerPrecedenceSpec);
-      const operation = oas.operation('/root-server', 'get');
-
-      expect(operation.getServers()).toStrictEqual(operationServerPrecedenceSpec.servers);
-      expect(operation.url()).toBe('https://us.root.example.com/v1');
-    });
-
-    it('should prefer path-item servers over root-level servers', () => {
-      const oas = Oas.init(operationServerPrecedenceSpec);
-      const operation = oas.operation('/path-server', 'get');
-
-      expect(operation.getServers()).toStrictEqual([
-        {
-          url: 'https://{region}.path.example.com/{version}',
-          variables: {
-            region: { default: 'eu', enum: ['eu', 'us'] },
-            version: { default: 'v2' },
-          },
-        },
-      ]);
-      expect(operation.url()).toBe('https://eu.path.example.com/v2');
-      expect(operation.defaultVariables()).toStrictEqual({ region: 'eu', version: 'v2' });
-      expect(operation.splitUrl()[1]).toStrictEqual({
-        key: 'region-1',
-        type: 'variable',
-        value: 'region',
-        description: undefined,
-        enum: ['eu', 'us'],
-      });
-    });
-
-    it('should prefer operation servers over path-item and root-level servers', () => {
-      const oas = Oas.init(operationServerPrecedenceSpec);
-      const operation = oas.operation('/operation-server', 'get');
-
-      expect(operation.getServers()).toStrictEqual([
-        {
-          url: 'https://{region}.operation.example.com/{version}',
-          variables: {
-            region: { default: 'apac', enum: ['apac', 'emea'] },
-            version: { default: 'v3' },
-          },
-        },
-      ]);
-      expect(operation.url()).toBe('https://apac.operation.example.com/v3');
-      expect(operation.variables()).toStrictEqual({
-        region: { default: 'apac', enum: ['apac', 'emea'] },
-        version: { default: 'v3' },
-      });
-    });
-
-    it('should embellish operation server defaults with user variables', () => {
-      const oas = new Oas(operationServerPrecedenceSpec, { region: 'emea' });
-      const operation = oas.operation('/operation-server', 'get');
-
-      expect(operation.defaultVariables()).toStrictEqual({ region: 'emea', version: 'v3' });
-      expect(operation.url()).toBe('https://emea.operation.example.com/v3');
-    });
-
-    it('should use the selected operation server and its variables', () => {
-      const oas = Oas.init(operationServerPrecedenceSpec);
-      const operation = oas.operation('/multiple-operation-servers', 'get');
-
-      expect(operation.url()).toBe('https://first.operation.example.com');
-      expect(operation.url(1)).toBe('https://us.second.operation.example.com/v4');
-      expect(operation.defaultVariables(1)).toStrictEqual({ region: 'us', version: 'v4' });
-    });
-
-    it('should normalize relative path-item and operation servers', () => {
-      const oas = Oas.init(operationServerPrecedenceSpec);
-
-      expect(oas.operation('/relative-path-server', 'get').url()).toBe('https://example.com/v2');
-      expect(oas.operation('/relative-operation-server', 'get').url()).toBe('https://example.com/v3');
-    });
-
-    it('should resolve path-item refs when retrieving path-item servers', () => {
-      const oas = Oas.init(operationServerPrecedenceSpec);
-      const operation = oas.operation('/path-item-ref-server', 'get');
-
-      expect(operation.getServers()).toStrictEqual([{ url: 'https://path-item-ref.example.com' }]);
-      expect(operation.url()).toBe('https://path-item-ref.example.com');
-    });
-
-    it('should ignore empty operation and path-item server arrays when applying precedence', () => {
-      const oas = Oas.init(operationServerPrecedenceSpec);
-
-      expect(oas.operation('/empty-operation-servers', 'get').url()).toBe('https://empty-operation-path.example.com');
-      expect(oas.operation('/empty-path-item-servers', 'get').url()).toBe('https://us.root.example.com/v1');
-    });
-
-    it('should return the default server if no servers are available', () => {
-      const oas = Oas.init({
-        openapi: '3.0.0',
-        info: { title: 'no servers', version: '1.0.0' },
-        paths: {
-          '/endpoint': {
-            get: {
-              responses: { 200: { description: 'OK' } },
-            },
-          },
-        },
-      });
-      const operation = oas.operation('/endpoint', 'get');
-
-      expect(operation.getServers()).toStrictEqual([{ url: 'https://example.com' }]);
-      expect(operation.url()).toBe('https://example.com');
-    });
-
-    it('should fall back to the default server URL when the selected server is unavailable', () => {
-      const oas = Oas.init(operationServerPrecedenceSpec);
-      const operation = oas.operation('/operation-server', 'get');
-
-      expect(operation.url(1)).toBe('https://example.com');
-      expect(operation.url(99)).toBe('https://example.com');
-    });
-
-    it('should fall back to the default server URL when the selected server has no URL', () => {
-      const oas = Oas.init({
-        openapi: '3.0.0',
-        info: { title: 'malformed server', version: '1.0.0' },
-        paths: {
-          '/endpoint': {
-            get: {
-              servers: [{}],
-              responses: { 200: { description: 'OK' } },
-            },
-          },
-        },
-      });
-      const operation = oas.operation('/endpoint', 'get');
-
-      expect(operation.getServers()).toStrictEqual([{}]);
-      expect(operation.url()).toBe('https://example.com');
     });
   });
